+++ src/i18n-gs.config.template.js
@@ -1,0 +1,17 @@
+module.exports = {
+  locales: {
+    // include: [],
+  },
+  sheets: {
+    // include: [],
+  },
+  credential: {
+    // path: "",
+  },
+  logging: {
+    level: "info",
+  },
+  i18n: {
+    keyStyle: "auto" | "flat" | "nested",
+  },
+};